Cars counting to examples
