+++ example/CarsCounting.cpp
@@ -5,13 +5,7 @@
 /// \param parser
 ///
 CarsCounting::CarsCounting(const cv::CommandLineParser& parser)
-    :
-      m_showLogs(true),
-      m_fps(25),
-      m_isTrackerInitialized(false),
-      m_startFrame(0),
-      m_endFrame(0),
-      m_finishDelay(0)
+    : VideoExample(parser)
 {
 #ifdef _WIN32
 	std::string pathToModel = "../../data/";
@@ -19,13 +13,6 @@ CarsCounting::CarsCounting(const cv::CommandLineParser& parser)
 	std::string pathToModel = "../data/";
 #endif
 
-
-    m_inFile = parser.get<std::string>(0);
-    m_outFile = parser.get<std::string>("out");
-    m_showLogs = parser.get<int>("show_logs") != 0;
-    m_startFrame = parser.get<int>("start_frame");
-    m_endFrame = parser.get<int>("end_frame");
-    m_finishDelay = parser.get<int>("end_delay");
 	m_drawHeatMap = parser.get<int>("heat_map") != 0;
 
 	m_weightsFile = parser.get<std::string>("weights");
@@ -50,178 +37,36 @@ CarsCounting::CarsCounting(const cv::CommandLineParser& parser)
 		m_detectorType = tracking::Detectors::Yolo_Darknet;
 
 	std::cout << "Inference framework set " << inference << " used " << m_detectorType << ", weights: " << m_weightsFile << ", config: " << m_configFile << ", names: " << m_namesFile << std::endl;
-
-    m_colors.emplace_back(255, 0, 0);
-    m_colors.emplace_back(0, 255, 0);
-    m_colors.emplace_back(0, 0, 255);
-    m_colors.emplace_back(255, 255, 0);
-    m_colors.emplace_back(0, 255, 255);
-    m_colors.emplace_back(255, 0, 255);
-    m_colors.emplace_back(255, 127, 255);
-    m_colors.emplace_back(127, 0, 255);
-    m_colors.emplace_back(127, 0, 127);
-}
-
-///
-/// \brief CarsCounting::~CarsCounting
-///
-CarsCounting::~CarsCounting()
-{
-
-}
-
-///
-/// \brief CarsCounting::Process
-///
-void CarsCounting::Process()
-{
-    cv::VideoWriter writer;
-
-#ifndef SILENT_WORK
-    cv::namedWindow("Video", cv::WINDOW_NORMAL | cv::WINDOW_KEEPRATIO);
-#endif
-
-    int k = 0;
-
-    double freq = cv::getTickFrequency();
-
-    int64 allTime = 0;
-
-    bool manualMode = false;
-    int framesCounter = m_startFrame + 1;
-
-    cv::VideoCapture capture;
-    if (m_inFile.size() == 1)
-        capture.open(atoi(m_inFile.c_str()));
-    else
-        capture.open(m_inFile);
-
-    if (!capture.isOpened())
-    {
-        std::cerr << "Can't open " << m_inFile << std::endl;
-        return;
-    }
-    capture.set(cv::CAP_PROP_POS_FRAMES, m_startFrame);
-
-    m_fps = std::max(1.f, (float)capture.get(cv::CAP_PROP_FPS));
-
-    cv::Mat colorFrame;
-	capture >> colorFrame;
-	if (colorFrame.empty())
-	{
-		std::cerr << "Frame is empty!" << std::endl;
-		return;
-	}
-	if (!m_isTrackerInitialized)
-	{
-		cv::UMat uframe = colorFrame.getUMat(cv::ACCESS_READ);
-		m_isTrackerInitialized = InitTracker(uframe);
-		if (!m_isTrackerInitialized)
-		{
-			std::cerr << "Tracker initialize error!!!" << std::endl;
-			return;
-		}
-	}
-
-    for (;;)
-    {
-        capture >> colorFrame;
-        if (colorFrame.empty())
-        {
-            std::cerr << "Frame is empty!" << std::endl;
-            break;
-        }
-
-        int64 t1 = cv::getTickCount();
-
-        cv::UMat uframe;
-        if (!m_detector->CanGrayProcessing() || m_tracker->CanColorFrameToTrack())
-            uframe = colorFrame.getUMat(cv::ACCESS_READ);
-		else
-			cv::cvtColor(colorFrame, uframe, cv::COLOR_BGR2GRAY);
-
-        m_detector->Detect(uframe);
-
-        const regions_t& regions = m_detector->GetDetects();
-
-        m_tracker->Update(regions, uframe, m_fps);
-
-        int64 t2 = cv::getTickCount();
-
-        allTime += t2 - t1;
-        int currTime = cvRound(1000 * (t2 - t1) / freq);
-
-        DrawData(colorFrame, framesCounter, currTime);
-
-#ifndef SILENT_WORK
-        cv::imshow("Video", colorFrame);
-
-		int waitTime = manualMode ? 0 : 1;// std::max<int>(1, cvRound(1000 / m_fps - currTime));
-        k = cv::waitKey(waitTime);
-        if (k == 'm' || k == 'M')
-            manualMode = !manualMode;
-        else if (k == 27)
-            break;
-
-#else
-		std::this_thread::sleep_for(std::chrono::milliseconds(1));
-#endif
-
-		if (!m_outFile.empty() && !writer.isOpened())
-			writer.open(m_outFile, cv::VideoWriter::fourcc('H', 'F', 'Y', 'U'), m_fps, colorFrame.size(), true);
-        if (writer.isOpened())
-            writer << colorFrame;
-
-        ++framesCounter;
-        if (m_endFrame && framesCounter > m_endFrame)
-        {
-            std::cout << "Process: riched last " << m_endFrame << " frame" << std::endl;
-            break;
-        }
-    }
-
-    std::cout << "work time = " << (allTime / freq) << std::endl;
-#ifndef SILENT_WORK
-	cv::waitKey(m_finishDelay);
-#endif
 }
 
 ///
 /// \brief CarsCounting::DrawTrack
 /// \param frame
-/// \param resizeCoeff
 /// \param track
 /// \param drawTrajectory
+/// \param framesCounters
 ///
-void CarsCounting::DrawTrack(cv::Mat frame,
-                             int resizeCoeff,
-                             const TrackingObject& track,
-                             bool drawTrajectory
-        )
+void CarsCounting::DrawTrack(cv::Mat frame, const TrackingObject& track, bool drawTrajectory, int framesCounter)
 {
-    auto ResizeRect = [&](const cv::Rect& r) -> cv::Rect
-    {
-        return cv::Rect(resizeCoeff * r.x, resizeCoeff * r.y, resizeCoeff * r.width, resizeCoeff * r.height);
-    };
-    auto ResizePoint = [&](const cv::Point& pt) -> cv::Point
-    {
-        return cv::Point(resizeCoeff * pt.x, resizeCoeff * pt.y);
-    };
+    cv::Rect brect = track.m_rrect.boundingRect();
+
+    m_resultsLog.AddTrack(framesCounter, track.m_ID, brect, track.m_type, track.m_confidence);
+	m_resultsLog.AddRobustTrack(track.m_ID);
 
     if (track.m_isStatic)
     {
 #if (CV_VERSION_MAJOR >= 4)
-        cv::rectangle(frame, ResizeRect(track.m_rrect.boundingRect()), cv::Scalar(255, 0, 255), 2, cv::LINE_AA);
+        cv::rectangle(frame, brect, cv::Scalar(255, 0, 255), 2, cv::LINE_AA);
 #else
-        cv::rectangle(frame, ResizeRect(track.m_rrect.boundingRect()), cv::Scalar(255, 0, 255), 2, CV_AA);
+        cv::rectangle(frame, brect, cv::Scalar(255, 0, 255), 2, CV_AA);
 #endif
     }
     else
     {
 #if (CV_VERSION_MAJOR >= 4)
-        cv::rectangle(frame, ResizeRect(track.m_rrect.boundingRect()), cv::Scalar(0, 255, 0), 1, cv::LINE_AA);
+        cv::rectangle(frame, brect, cv::Scalar(0, 255, 0), 1, cv::LINE_AA);
 #else
-        cv::rectangle(frame, ResizeRect(track.m_rrect.boundingRect()), cv::Scalar(0, 255, 0), 1, CV_AA);
+        cv::rectangle(frame, brect, cv::Scalar(0, 255, 0), 1, CV_AA);
 #endif
 
 		if (!m_geoParams.Empty())
@@ -246,7 +91,6 @@ void CarsCounting::DrawTrack(cv::Mat frame,
 				double fontScale = 0.5;
 				cv::Size labelSize = cv::getTextSize(label.str(), cv::FONT_HERSHEY_SIMPLEX, fontScale, 1, &baseLine);
 
-				cv::Rect brect = track.m_rrect.boundingRect();
 				if (brect.x < 0)
 				{
 					brect.width = std::min(brect.width, frame.cols - 1);
@@ -285,16 +129,16 @@ void CarsCounting::DrawTrack(cv::Mat frame,
             const TrajectoryPoint& pt1 = track.m_trace.at(j);
             const TrajectoryPoint& pt2 = track.m_trace.at(j + 1);
 #if (CV_VERSION_MAJOR >= 4)
-            cv::line(frame, ResizePoint(pt1.m_prediction), ResizePoint(pt2.m_prediction), cl, 1, cv::LINE_AA);
+            cv::line(frame, pt1.m_prediction, pt2.m_prediction, cl, 1, cv::LINE_AA);
 #else
-			cv::line(frame, ResizePoint(pt1.m_prediction), ResizePoint(pt2.m_prediction), cl, 1, CV_AA);
+			cv::line(frame, pt1.m_prediction, pt2.m_prediction, cl, 1, CV_AA);
 #endif
             if (!pt2.m_hasRaw)
             {
 #if (CV_VERSION_MAJOR >= 4)
-                cv::circle(frame, ResizePoint(pt2.m_prediction), 4, cl, 1, cv::LINE_AA);
+                cv::circle(frame, pt2.m_prediction, 4, cl, 1, cv::LINE_AA);
 #else
-				cv::circle(frame, ResizePoint(pt2.m_prediction), 4, cl, 1, CV_AA);
+				cv::circle(frame, pt2.m_prediction, 4, cl, 1, CV_AA);
 #endif
             }
         }
@@ -302,26 +146,11 @@ void CarsCounting::DrawTrack(cv::Mat frame,
 }
 
 ///
-/// \brief CarsCounting::InitTracker
-/// \param grayFrame
+/// \brief CarsCounting::InitDetector
+/// \param frame
 ///
-bool CarsCounting::InitTracker(cv::UMat frame)
+bool CarsCounting::InitDetector(cv::UMat frame)
 {
-	bool res = true;
-
-	if (m_drawHeatMap)
-	{
-		if (frame.channels() == 3)
-			m_keyFrame = frame.getMat(cv::ACCESS_READ).clone();
-		else
-			cv::cvtColor(frame, m_keyFrame, cv::COLOR_GRAY2BGR);
-		m_heatMap = cv::Mat(m_keyFrame.size(), CV_32FC1, cv::Scalar::all(0));
-	}
-
-    m_minObjWidth = frame.cols / 50;
-
-    const int minStaticTime = 5;
-
     config_t config;
 
 #if 1
@@ -378,68 +207,78 @@ bool CarsCounting::InitTracker(cv::UMat frame)
 
 #endif
 
-	if (m_detector.get())
-		m_detector->SetMinObjectSize(cv::Size(m_minObjWidth, m_minObjWidth));
-	else
-		res = false;
+	return m_detector.operator bool();
+}
 
-	if (res)
-	{
-		TrackerSettings settings;
-		settings.SetDistance(tracking::DistJaccard);
-		settings.m_kalmanType = tracking::KalmanLinear;
-		settings.m_filterGoal = tracking::FilterCenter;
-		settings.m_lostTrackType = tracking::TrackCSRT; // Use KCF tracker for collisions resolving. Used if m_filterGoal == tracking::FilterRect
-		settings.m_matchType = tracking::MatchHungrian;
-		settings.m_dt = 0.3f;                           // Delta time for Kalman filter
-		settings.m_accelNoiseMag = 0.2f;                // Accel noise magnitude for Kalman filter
-		settings.m_distThres = 0.7f;                    // Distance threshold between region and object on two frames
-		settings.m_minAreaRadiusPix = frame.rows / 20.f;
-		settings.m_maximumAllowedSkippedFrames = cvRound(2 * m_fps); // Maximum allowed skipped frames
+///
+/// \brief CarsCounting::InitTracker
+/// \param grayFrame
+///
+bool CarsCounting::InitTracker(cv::UMat frame)
+{
+	if (m_drawHeatMap)
+    {
+        if (frame.channels() == 3)
+            m_keyFrame = frame.getMat(cv::ACCESS_READ).clone();
+        else
+            cv::cvtColor(frame, m_keyFrame, cv::COLOR_GRAY2BGR);
+        m_heatMap = cv::Mat(m_keyFrame.size(), CV_32FC1, cv::Scalar::all(0));
+    }
 
-		settings.AddNearTypes(ObjectTypes::obj_car, ObjectTypes::obj_bus, false);
-		settings.AddNearTypes(ObjectTypes::obj_car, ObjectTypes::obj_truck, false);
-		settings.AddNearTypes(ObjectTypes::obj_person, ObjectTypes::obj_bicycle, true);
-		settings.AddNearTypes(ObjectTypes::obj_person, ObjectTypes::obj_motorbike, true);
+    const int minStaticTime = 5;
 
-		settings.m_useAbandonedDetection = false;
-		if (settings.m_useAbandonedDetection)
-		{
-			settings.m_minStaticTime = minStaticTime;
-			settings.m_maxStaticTime = 60;
-			settings.m_maximumAllowedSkippedFrames = cvRound(settings.m_minStaticTime * m_fps); // Maximum allowed skipped frames
-			settings.m_maxTraceLength = 2 * settings.m_maximumAllowedSkippedFrames;        // Maximum trace length
-		}
-		else
-		{
-			settings.m_maximumAllowedSkippedFrames = cvRound(10 * m_fps); // Maximum allowed skipped frames
-			settings.m_maxTraceLength = cvRound(4 * m_fps);              // Maximum trace length
-		}
+    TrackerSettings settings;
+    settings.SetDistance(tracking::DistJaccard);
+    settings.m_kalmanType = tracking::KalmanLinear;
+    settings.m_filterGoal = tracking::FilterCenter;
+    settings.m_lostTrackType = tracking::TrackCSRT; // Use KCF tracker for collisions resolving. Used if m_filterGoal == tracking::FilterRect
+    settings.m_matchType = tracking::MatchHungrian;
+    settings.m_dt = 0.3f;                           // Delta time for Kalman filter
+    settings.m_accelNoiseMag = 0.2f;                // Accel noise magnitude for Kalman filter
+    settings.m_distThres = 0.7f;                    // Distance threshold between region and object on two frames
+    settings.m_minAreaRadiusPix = frame.rows / 20.f;
+    settings.m_maximumAllowedSkippedFrames = cvRound(2 * m_fps); // Maximum allowed skipped frames
 
-		m_tracker = BaseTracker::CreateTracker(settings);
-	}
+    settings.AddNearTypes(ObjectTypes::obj_car, ObjectTypes::obj_bus, false);
+    settings.AddNearTypes(ObjectTypes::obj_car, ObjectTypes::obj_truck, false);
+    settings.AddNearTypes(ObjectTypes::obj_person, ObjectTypes::obj_bicycle, true);
+    settings.AddNearTypes(ObjectTypes::obj_person, ObjectTypes::obj_motorbike, true);
+
+    settings.m_useAbandonedDetection = false;
+    if (settings.m_useAbandonedDetection)
+    {
+        settings.m_minStaticTime = minStaticTime;
+        settings.m_maxStaticTime = 60;
+        settings.m_maximumAllowedSkippedFrames = cvRound(settings.m_minStaticTime * m_fps); // Maximum allowed skipped frames
+        settings.m_maxTraceLength = 2 * settings.m_maximumAllowedSkippedFrames;        // Maximum trace length
+    }
+    else
+    {
+        settings.m_maximumAllowedSkippedFrames = cvRound(10 * m_fps); // Maximum allowed skipped frames
+        settings.m_maxTraceLength = cvRound(4 * m_fps);              // Maximum trace length
+    }
+
+    m_tracker = BaseTracker::CreateTracker(settings);
+
 #if 0
 #if 0
-	std::vector<cv::Point> framePoints{ cv::Point(420, 348), cv::Point(509, 283), cv::Point(731, 281), cv::Point(840, 343) };
-	std::vector<cv::Point2f> geoPoints{ cv::Point2f(45.526646, 5.974535), cv::Point2f(45.527566, 5.973849), cv::Point2f(45.527904, 5.974135), cv::Point2f(45.526867, 5.974826) };
+    std::vector<cv::Point> framePoints{ cv::Point(420, 348), cv::Point(509, 283), cv::Point(731, 281), cv::Point(840, 343) };
+    std::vector<cv::Point2f> geoPoints{ cv::Point2f(45.526646, 5.974535), cv::Point2f(45.527566, 5.973849), cv::Point2f(45.527904, 5.974135), cv::Point2f(45.526867, 5.974826) };
 #else
-	std::vector<cv::Point> framePoints{ cv::Point(1665, 746), cv::Point(246, 521), cv::Point(570, 282), cv::Point(1773, 378) };
-	std::vector<cv::Point2f> geoPoints{ cv::Point2f(30.258855, 60.006536), cv::Point2f(30.258051, 60.006855), cv::Point2f(30.258080, 60.007414), cv::Point2f(30.259066, 60.007064) };
+    std::vector<cv::Point> framePoints{ cv::Point(1665, 746), cv::Point(246, 521), cv::Point(570, 282), cv::Point(1773, 378) };
+    std::vector<cv::Point2f> geoPoints{ cv::Point2f(30.258855, 60.006536), cv::Point2f(30.258051, 60.006855), cv::Point2f(30.258080, 60.007414), cv::Point2f(30.259066, 60.007064) };
 #endif
-	m_geoParams.SetKeyPoints(framePoints, geoPoints);
+    m_geoParams.SetKeyPoints(framePoints, geoPoints);
 #endif
-    return res;
+    return true;
 }
 
 ///
 /// \brief CarsCounting::DrawData
 /// \param frame
 ///
-void CarsCounting::DrawData(cv::Mat frame, int framesCounter, int currTime)
+void CarsCounting::DrawData(cv::Mat frame, const std::vector<TrackingObject>& tracks, int framesCounter, int currTime)
 {
-	std::vector<TrackingObject> tracks;
-	m_tracker->GetTracks(tracks);
-
     if (m_showLogs)
         std::cout << "Frame " << framesCounter << ": tracks = " << tracks.size() << ", time = " << currTime << std::endl;
 
@@ -456,7 +295,7 @@ void CarsCounting::DrawData(cv::Mat frame, int framesCounter, int currTime)
     {
         if (track.m_isStatic)
         {
-            DrawTrack(frame, 1, track, true);
+            DrawTrack(frame, track, true, framesCounter);
         }
         else
         {
@@ -465,7 +304,7 @@ void CarsCounting::DrawData(cv::Mat frame, int framesCounter, int currTime)
                                 cv::Size2f(0.1f, 8.0f))      // Min and max ratio: width / height
                     )
             {
-                DrawTrack(frame, 1, track, true);
+                DrawTrack(frame, track, true, framesCounter);
 
                 CheckLinesIntersection(track, static_cast<float>(frame.cols), static_cast<float>(frame.rows));
             }
