+++ example/VideoExample.h
@@ -1,9 +1,7 @@
 #pragma once
 
 #include <iostream>
-#include <fstream>
 #include <vector>
-#include <map>
 #include <thread>
 #include <mutex>
 #include <chrono>
@@ -12,157 +10,7 @@
 
 #include "BaseDetector.h"
 #include "Ctracker.h"
-
-///
-/// \brief The ResultsLog class
-///
-class ResultsLog
-{
-public:
-	///
-	ResultsLog(const std::string& fileName)
-		: m_fileName(fileName)
-	{
-	}
-
-	///
-	~ResultsLog()
-	{
-		WriteAll(true);
-	}
-
-	///
-	bool Open()
-	{
-		m_resCSV.close();
-		if (m_fileName.size() > 5)
-		{
-			m_resCSV.open(m_fileName);
-			return m_resCSV.is_open();
-		}
-		return false;
-	}
-
-	///
-	bool AddTrack(int framesCounter, size_t trackID, const cv::Rect& brect, objtype_t type, float confidence)
-	{
-		if (m_resCSV.is_open())
-		{
-			auto frame = m_frames.find(framesCounter);
-			if (frame == std::end(m_frames))
-			{
-				DetectsOnFrame tmpFrame;
-				tmpFrame.m_detects.emplace_back(trackID, brect, type, confidence);
-				m_frames.emplace(framesCounter, tmpFrame);
-			}
-			else
-			{
-				frame->second.m_detects.emplace_back(trackID, brect, type, confidence);
-			}
-			return true;
-		}
-		return false;
-	}
-
-	///
-	void AddRobustTrack(size_t trackID)
-	{
-		m_robustIDs.insert(trackID);
-	}
-
-    ///
-    void Flush()
-    {
-        WriteAll(true);
-        m_frames.clear();
-    }
-
-private:
-	std::string m_fileName;
-	std::ofstream m_resCSV;
-
-	///
-	struct Detection
-	{
-		cv::Rect m_rect;
-		objtype_t m_type;
-		float m_conf = 0.f;
-		size_t m_trackID = 0;
-
-		Detection(size_t trackID, const cv::Rect& brect, objtype_t type, float confidence)
-		{
-			m_type = type;
-			m_rect = brect;
-			m_conf = confidence;
-			m_trackID = trackID;
-		}
-	};
-
-	///
-	struct DetectsOnFrame
-	{
-		std::vector<Detection> m_detects;
-	};
-	std::map<int, DetectsOnFrame> m_frames;
-	std::set<size_t> m_robustIDs;
-
-	///
-	void WriteAll(bool byFrames)
-	{
-		if (byFrames)
-		{
-#if 1
-			char delim = ',';
-			for (const auto& frame : m_frames)
-			{
-				for (const auto& detect : frame.second.m_detects)
-				{
-					if (m_robustIDs.find(detect.m_trackID) != std::end(m_robustIDs))
-					{
-						m_resCSV << frame.first << delim << TypeConverter::Type2Str(detect.m_type) << delim << detect.m_rect.x << delim << detect.m_rect.y << delim <<
-							detect.m_rect.width << delim << detect.m_rect.height << delim <<
-							detect.m_conf << delim << std::endl;
-					}
-				}
-			}
-#else
-			char delim = '	';
-			for (const auto& frame : m_frames)
-			{
-				for (const auto& detect : frame.second.m_detects)
-				{
-					if (m_robustIDs.find(detect.m_trackID) != std::end(m_robustIDs))
-					{
-						m_resCSV << frame.first << delim << TypeConverter::Type2Str(detect.m_type) << delim << detect.m_rect.x << delim << detect.m_rect.y << delim <<
-							(detect.m_rect.x + detect.m_rect.width) << delim << (detect.m_rect.y + detect.m_rect.height) << delim <<
-							detect.m_conf << delim << detect.m_trackID << std::endl;
-					}
-				}
-			}
-#endif
-		}
-		else
-		{
-			char delim = ',';
-			for (size_t id : m_robustIDs)
-			{
-				for (const auto& frame : m_frames)
-				{
-					for (const auto& detect : frame.second.m_detects)
-					{
-						if (detect.m_trackID == id)
-						{
-							m_resCSV << frame.first << delim << id << delim << detect.m_rect.x << delim << detect.m_rect.y << delim <<
-								detect.m_rect.width << delim << detect.m_rect.height << delim <<
-								detect.m_conf << ",-1,-1,-1," << std::endl;
-							break;
-						}
-					}
-				}
-			}
-		}
-	}
-};
+#include "FileLogger.h"
 
 ///
 /// \brief The Frame struct
@@ -365,10 +213,12 @@ protected:
 
     virtual void DrawData(cv::Mat frame, const std::vector<TrackingObject>& tracks, int framesCounter, int currTime) = 0;
 
-    void DrawTrack(cv::Mat frame, int resizeCoeff, const TrackingObject& track, bool drawTrajectory, int framesCounter);
+    virtual void DrawTrack(cv::Mat frame, const TrackingObject& track, bool drawTrajectory, int framesCounter);
 
     TrackerSettings m_trackerSettings;
     bool m_trackerSettingsLoaded = false;
+
+    std::vector<cv::Scalar> m_colors;
 
 private:
 	std::vector<TrackingObject> m_tracks;
@@ -381,7 +231,6 @@ private:
     int m_startFrame = 0;
     int m_endFrame = 0;
     int m_finishDelay = 0;
-    std::vector<cv::Scalar> m_colors;
 
     FrameInfo m_frameInfo[2];
