+++ example/examples.h
@@ -165,7 +165,7 @@ protected:
         {
             if (track.m_isStatic)
             {
-                DrawTrack(frame, 1, track, false, framesCounter);
+                DrawTrack(frame, track, false, framesCounter);
 
 				std::string label = "abandoned " + std::to_string(track.m_ID);
 				int baseLine = 0;
@@ -200,7 +200,7 @@ protected:
                 if (track.IsRobust(cvRound(m_fps / 4),          // Minimal trajectory size
                                     0.7f,                        // Minimal ratio raw_trajectory_points / trajectory_lenght
                                     cv::Size2f(0.1f, 8.0f)))      // Min and max ratio: width / height
-                    DrawTrack(frame, 1, track, true, framesCounter);
+                    DrawTrack(frame, track, true, framesCounter);
             }
         }
         m_detector->CalcMotionMap(frame);
@@ -291,7 +291,7 @@ protected:
             if (track.IsRobust(8,                           // Minimal trajectory size
                                 0.4f,                        // Minimal ratio raw_trajectory_points / trajectory_lenght
                                 cv::Size2f(0.1f, 8.0f)))      // Min and max ratio: width / height
-                DrawTrack(frame, 1, track, true, framesCounter);
+                DrawTrack(frame, track, true, framesCounter);
         }
         m_detector->CalcMotionMap(frame);
     }
@@ -381,7 +381,7 @@ protected:
 			if (track.IsRobust(cvRound(m_fps / 2),          // Minimal trajectory size
                                 0.4f,                        // Minimal ratio raw_trajectory_points / trajectory_lenght
                                 cv::Size2f(0.1f, 8.0f)))      // Min and max ratio: width / height
-                DrawTrack(frame, 1, track, true, framesCounter);
+                DrawTrack(frame, track, true, framesCounter);
         }
         m_detector->CalcMotionMap(frame);
     }
@@ -521,7 +521,7 @@ protected:
 				0.5f,                        // Minimal ratio raw_trajectory_points / trajectory_lenght
 				cv::Size2f(0.1f, 8.0f)))      // Min and max ratio: width / height
 			{
-				DrawTrack(frame, 1, track, false, framesCounter);
+				DrawTrack(frame, track, false, framesCounter);
 
 
 				std::stringstream label;
@@ -761,7 +761,7 @@ protected:
                 0.5f,                        // Minimal ratio raw_trajectory_points / trajectory_lenght
 				cv::Size2f(0.1f, 8.0f)))      // Min and max ratio: width / height
 			{
-				DrawTrack(frame, 1, track, false, framesCounter);
+				DrawTrack(frame, track, false, framesCounter);
 
 
 				std::stringstream label;
@@ -979,7 +979,7 @@ protected:
 				0.5f,                        // Minimal ratio raw_trajectory_points / trajectory_lenght
 				cv::Size2f(0.1f, 8.0f)))      // Min and max ratio: width / height
 			{
-				DrawTrack(frame, 1, track, true, framesCounter);
+				DrawTrack(frame, track, true, framesCounter);
 
 
 				std::stringstream label;
